+++ src/cpp/tests/automation/testthat/test-automation-completions.R
@@ -1,0 +1,29 @@
+
+library(testthat)
+
+self <- remote <- .rs.automation.newRemote()
+on.exit(.rs.automation.deleteRemote(), add = TRUE)
+
+# https://github.com/rstudio/rstudio/issues/14784
+test_that("autocompletion doesn't trigger active bindings", {
+   
+   code <- .rs.heredoc(r'{
+      Test <- R6::R6Class("Test",
+        active  = list(active_test  = function(value) print("active")),
+        private = list(private_test = function(value) print("private")),
+        public  = list(public_test  = function(value) print("public"))
+      )
+      
+      n <- Test$new()
+      nms <- .rs.getNames(n)
+   }')
+   
+   remote$consoleExecute(code)
+   output <- remote$consoleOutput()
+   expect_false(tail(output, 1) == "[1] \"active\"")
+   
+   remote$keyboardExecute("n", "<Tab>", "<Escape>", "<Backspace>")
+   output <- remote$consoleOutput()
+   expect_false(tail(output, 1) == "[1] \"active\"")
+   
+})